Remove StatusesError, log directly when desired
The intend was to look for StatusesError in our overall catch and log
the statuses it contains. This would ensure we always see the final
state when we're crashing. For example, if a required status had failed,
this is how we'd show it.

This hasn't been working. The `instanceof StatusesError` seems to return
`false` and we fall through to just printing the `message`. This makes
it annoying to figure out the required-status-failed case because we
never see it.

Rather than debug JavaScript's object system, as implemented on GitHub
actions, I've instead changed things to throw a simple Error and
manually log the statuses before throwing it. This should be more
reliable, and it gives us the opportunity to only log the failing
statuses for that particular error, which seems nice.